+++ src/sound/5220intf.h
@@ -29,7 +29,7 @@ double tms5220_time_to_ready(void);
 int tms5220_int_r(void);
 
 void tms5220_reset(void);
-void tms5220_set_frequency(int frequency);
+void tms5220_set_frequency(double frequency);
 
 #endif
